% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cremisc_func.R
\name{from_excel}
\alias{from_excel}
\title{Copy tabular data from excel}
\usage{
from_excel()
}
\description{
\code{from_excel} reads tabular data copied from excel,
}
\details{
This function makes it easier to transfer data from excel into R.
}

