% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cremisc_func.R
\name{load_packages}
\alias{load_packages}
\title{Load packages}
\usage{
load_packages(...)
}
\arguments{
\item{...}{sequence of packages to load}
}
\description{
\code{loadPackages} loads packages and installs it if needed.
This function is meant to replace the tedious task of loading packages and
loading them individual. Here, everything is handled in a single step.
}
\examples{
loadPackages(ape, xtable)
}

