% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cremisc_func.R
\name{loadPackages}
\alias{loadPackages}
\title{Load packages}
\usage{
loadPackages(...)
}
\arguments{
\item{...}{sequence of packages to load}
}
\description{
\code{loadPackages} loads packages and installs it if needed.
}
\details{
This function is meant to replace the tedious task of loading packages and
loading them individual. Here, everything is handled in a single step.
}
\examples{
loadPackages(ape, xtable)

}

