% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cremisc_func.R
\name{to_excel}
\alias{to_excel}
\title{Copy tabular data from excel}
\usage{
to_excel(x)
}
\description{
\code{to_excel} copies a data.frame to the clipboard, ready to be copied into excel
}
\details{
This function makes it easier to transfer data from R into excel.
}

